requirements.txt
html5lib
requests
dateparser
calmjs.parse
beautifulsoup4